Move organizer indexation service to separate service provider
