+++ app/Organizer/OrganizerServiceProvider.php
@@ -58,47 +58,5 @@ final class OrganizerServiceProvider extends BaseServiceProvider
                 );
             }
         );
-
-        $this->add(
-            'organizer_search_projector',
-            function () {
-                $service = new TransformingJsonDocumentIndexService(
-                    $this->get(JsonDocumentFetcher::class),
-                    $this->get('organizer_elasticsearch_transformer'),
-                    $this->get('organizer_elasticsearch_repository')
-                );
-
-                $service->setLogger($this->get('logger.amqp.udb3_consumer'));
-
-                return new OrganizerSearchProjector($service);
-            },
-            'event_bus_subscribers'
-        );
-
-        $this->add(
-            'organizer_elasticsearch_transformer',
-            function () {
-                return new JsonDocumentTransformer(
-                    new OrganizerTransformer(
-                        new JsonTransformerPsrLogger(
-                            $this->get('logger.amqp.udb3_consumer')
-                        ),
-                        new PathEndIdUrlParser()
-                    )
-                );
-            }
-        );
-
-        $this->add(
-            'organizer_elasticsearch_repository',
-            function () {
-                return new ElasticSearchDocumentRepository(
-                    $this->get(Client::class),
-                    $this->parameter('elasticsearch.organizer.write_index'),
-                    $this->parameter('elasticsearch.organizer.document_type'),
-                    $this->get('elasticsearch_indexation_strategy')
-                );
-            }
-        );
     }
 }
